Change trans ID to 64-bit to accommodate GSAP HRefs

+++ Pod/Classes/Terminals/PAX/Builders/HpsPaxCreditAdjustBuilder.h
@@ -16,11 +16,11 @@
     HpsPaxDevice *device;
 }
 
-@property (nonatomic, readwrite) int referenceNumber;
+@property (nonatomic, readwrite) NSInteger referenceNumber;
 @property (nonatomic, strong) NSNumber *amount;
 @property (nonatomic, strong) NSNumber *gratuity;
-@property (nonatomic, readwrite) int transactionId;
-@property (nonatomic, readwrite) int transactionNumber;
+@property (nonatomic, readwrite) NSInteger transactionId;
+@property (nonatomic, readwrite) NSInteger transactionNumber;
 
 - (void) execute:(void(^)(HpsPaxCreditResponse*, NSError*))responseBlock;
 - (id)initWithDevice: (HpsPaxDevice*)paxDevice;
