+++ Pod/Classes/Terminals/PAX/Builders/HpsPaxCreditVoidBuilder.h
@@ -16,9 +16,9 @@
     HpsPaxDevice *device;
 }
 
-@property (nonatomic, readwrite) int referenceNumber;
-@property (nonatomic, readwrite) int transactionId;
-@property (nonatomic, readwrite) int transactionNumber;
+@property (nonatomic, readwrite) NSInteger referenceNumber;
+@property (nonatomic, readwrite) NSInteger transactionId;
+@property (nonatomic, readwrite) NSInteger transactionNumber;
 
 - (void) execute:(void(^)(HpsPaxCreditResponse*, NSError*))responseBlock;
 - (id)initWithDevice: (HpsPaxDevice*)paxDevice;
